refactor: Aside & Header 분리
