+++ src/assets/article/index.ts
@@ -1,0 +1,5 @@
+export const INFO = {
+  NAME: '양이진 Yang Leejin',
+  INTRO:
+    '오래전에 함께 듣던 노래가 발걸음을 다시 멈춰서게 해 이 거리에서 너를 느낄 수 있어 널 이곳에서 꼭 다시 만날 것 같아 너일까 봐 한 번 더 바라보고 너일까 봐 자꾸 돌아보게 돼 어디선가 같은 노래를 듣고 날 생각하며 너 역시 멈춰있을까 오래전에 함께 듣던 노래가 거리에서 내게 우연히 들려온 것처럼 살아가다 한 번쯤 우연히 만날 것 같아 사랑했던 그 모습 그대로 오래전에 함께 듣던 노래가 발걸음을 다시 멈춰서게 해 이 거리에서 너를 느낄 수 있어 널 이곳에서 꼭 다시 만날 것 같아 너일까 봐 한 번 더 바라보고 너일까 봐 자꾸 돌아보게 돼 어디선가 같은 노래를 듣고 날 생각하며 너 역시 멈춰있을까 오래전에 함께 듣던 노래가 거리에서 내게 우연히 들려온 것처럼 살아가다 한 번쯤 우연히 만날 것 같아 사랑했던 그 모습',
+};
