+++ src/routes/Portfolio/Header/header.module.scss
@@ -1,0 +1,34 @@
+@use '/src/styles/constants/colors';
+@use '/src/styles/mixins/responsive';
+
+.header {
+  position: fixed;
+  top: 0 !important;
+  z-index: 2;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+  height: 60px;
+  padding: 0 20px;
+  color: var(--color-text);
+  background: var(--primary);
+  transition: background 0.2s ease-in 0s;
+
+  .nav {
+    display: flex;
+
+    li {
+      padding: 0 4px;
+      font-size: 14px;
+    }
+
+    li:last-of-type {
+      padding: 0 0 0 4px;
+    }
+  }
+
+  @include responsive.after(SD) {
+    display: none;
+  }
+}
