+++ src/styles/base/_fonts.scss
@@ -6,7 +6,6 @@ button,
 input,
 textarea {
   font-family: 'Spoqa Han Sans Neo', Inter, Arial, PingFangSC-Regular, 'Microsoft YaHei', sans-serif;
-  font-size: 16px;
   font-feature-settings: 'tnum';
 }
