+++ src/styles/base/_reset.scss
@@ -90,6 +90,7 @@ video {
   font-size: 100%;
   vertical-align: baseline;
   border: 0;
+  text-size-adjust: none;
 }
 /* HTML5 display-role reset for older browsers */
 article,
