+++ src/styles/base/_theme.scss
@@ -66,7 +66,7 @@
 }
 
 :root[color-theme='green'] {
-  --primary: #004d3c;
+  --primary: #004031;
   --color-text: #ececec;
 }
