+++ src/styles/constants/_colors.scss
@@ -11,7 +11,7 @@ $BORDER_DIM: #e0e0e0;
 $BORDER_DARK: #e0e0e0;
 
 $PRIMARY_ORANGE: #c6643a;
-$PRIMARY_GREEN: #004d3c;
+$PRIMARY_GREEN: #004031;
 $PRIMARY_BLUE: #809ea8;
 $PRIMARY_YELLOW: #e29f04;
 $PRIMARY_PINK: #e08e70;
